Concatenation example
Time Traveler!! You got a new time traveler - rock on! It needs a little programming to set up. You must write a program asking for the current time (hours only) and an amount of hours in the future. Use the modulo % operator to tell the time traveler the future hour to which they will be traveling.

Use a 24-hour clock and do not worry about AM/PM. For instance, if the current time is 20 and it is 6 hours in the future, it would be 2.

RUNSAVELOADSHOW CODELENS