Checking for user age
Are you legal? This program asks the user for their age, then outputs the things this person is legally able to do. As a reminder, at 16 you are allowed to drive, at 18 you are allowed to vote, and at 21 you are allowed to drink. If the person is under 16, your program should print: Some day, kid...