Donuts
Donut worry, be happy! The donut shop offers discounts on their surprise dozen special on even-numbered Sundays (for example, Sunday the 24th and not Sunday the 17th). Normally, donuts cost 99 cents a piece and there is no discount for buying a dozen. On even-numbered Sundays, you get 25% off the total price if you buy a dozen. Write a program that calculates the price of a dozen donuts on a given day.