concerts june
TGIF, finally! On Friday nights during June, the local art museum hosts a concert. If it is not raining, the concert will be held outside in the adjacent park. If it is raining, the concert will be held in the museum. Create a program that asks the user for the month, the day of the week, and the weather (rain or sun) then outputs whether if there will be a concert, and if so where it will be held.

Hint: First, try creating the program without worrying about the weather. Then try adding the weather.

> What month is it? June
> What day is it? Tuesday
> Is it rainy or sunny? rainy
Sorry, concerts are only held on Fridays!
> What month is it? December
> What day is it? Friday
> Is it rainy or sunny? sunny
Sorry, concerts are only held during June!
> What month is it? June
> What day is it? Friday
> Is it rainy or sunny? rain
There is a concert tonight! It will be held inside the art museum.