combining_concepts
We know that we can determine whether an integer is even or odd by the modulus operator (%). And we’ve seen several examples where decided whether or not to print a message based on whether a given number was even or odd, using a conditional. Let’s combine this concept with that of a loop to print a list of even numbers.